+++ src/workspace.ts
@@ -141,12 +141,16 @@ export {
     SerializedLayoutLink, SerializedLayoutLinkGroup, SerializedLayoutLinkItem,
 } from './editor/serializedDiagram';
 
-export { ClassicTemplate } from './templates/classicTemplate';
+export { ClassicTemplate, ClassicEntity, ClassicEntityProps } from './templates/classicTemplate';
 export {
     DefaultLinkTemplate, DefaultLinkPathTemplate, DefaultLinkPathTemplateProps,
 } from './templates/defaultLinkTemplate';
 export { GroupPaginator, GroupPaginatorProps } from './templates/groupPaginator';
-export { StandardTemplate, StandardEntity, StandardEntityProps } from './templates/standardTemplate';
+export { RoundTemplate, RoundEntity, RoundEntityProps } from './templates/roundTemplate';
+export {
+    StandardTemplate, StandardEntity, StandardEntityProps,
+    StandardEntityGroup, StandardEntityGroupProps,
+} from './templates/standardTemplate';
 export { SemanticTypeStyles } from './templates/typeStyles';
 export { OntologyLinkTemplates, LINK_STYLE_SHOW_IRI } from './templates/linkStyles';